fix(runner): checkpoint the whole-in-parts pass and the lint repairs

+++ internal/engine/retentionreport.go
@@ -350,7 +350,7 @@ func (r *retention) observed(out *statelog.Reading) {
 	// THE ANSWER COUNTERS FIRST, because two alarms are FRACTIONS of them
 	// and a fraction needs its denominator before either numerator means
 	// anything.
-	var answers, scoped, degraded uint64
+	var answers, scoped, degraded float64
 	for _, snapshot := range reading {
 		if snapshot.Name != metrics.TrackerSearchAnswers {
 			continue
@@ -364,8 +364,8 @@ func (r *retention) observed(out *statelog.Reading) {
 		}
 	}
 	if answers > 0 {
-		out.SearchScopedFraction = float64(scoped) / float64(answers)
-		out.SearchDegradedFraction = float64(degraded) / float64(answers)
+		out.SearchScopedFraction = scoped / answers
+		out.SearchDegradedFraction = degraded / answers
 	}
 
 	// THE DECLARED RATE, beside the observed one below. It is a constant
